# openapi.yaml
openapi: 3.0.3
info:
  title: Employee Attendance & Analytics API
  version: 2.0.0
  description: |
    Contract for the HROne engineering assignment. Your FastAPI app must follow this document exactly: paths, methods,
    parameter names, status codes and response field names. The application is organized into routers, schemas, services, configuration, and database modules;
    `app/main.py` exposes the FastAPI entry point and must start with `uvicorn app.main:app --port 8000`.

    **Where things are defined.** This file defines the HTTP contract and the business rules R1-R10. What is stored in
    MongoDB (collections, field types, what the grader seeds, nine sample documents) is defined in `DATA_MODEL.md`.

    ## Configuration
    `MONGO_URI` and `MONGO_DB` are read from environment variables (a local `.env` is fine; real environment variables
    must win). Create all indexes at startup, idempotently.

    ## Conventions used by every endpoint
    * **Employees are identified by `emp_code`.** It is a customer-generated business code assigned by the customer's HR
      team (this API does not generate it). It is unique, never changes, and is meant for humans. It is **not** MongoDB's
      `_id`. `_id` is internal: it never appears in a request or a response. Attendance records have no id either; a
      record is addressed by (`emp_code`, `date`).
    * **Instants are epoch milliseconds.** Every point in time in a request or response (`punched_at`, `punch_in`,
...

...

...

...

...

...

...

...
paths:
...
  /analytics/employees/{emp_code}/monthly:
    get:
      tags: [analytics]
      operationId: employeeMonthly
      summary: Monthly attendance summary for one employee
      description: |
        Backed by a MongoDB aggregation. `working_days` per R7. `present_days` per R5-R7 (half day = 0.5, Mon-Fri only).
        `leave_days` = number of `LEAVE` records. `late_count` = records with `late_minutes` > 0.
        `attendance_pct` = `present_days` / `working_days` x 100, R8 rounding, or `null` when `working_days` is 0.
        404 for an unknown employee.
      parameters:
        - {name: emp_code, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/Month'}
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EmployeeMonthly'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /analytics/departments/summary:
    get:
      tags: [analytics]
      operationId: departmentSummary
      summary: Department summary for a month
      description: |
        One item per department, sorted by `department` ascending, via a MongoDB aggregation across both collections.
        * `headcount` per R9 - **people with zero logs still count**, people who join later do not.
        * `present_days` as in the employee summary, summed over the department.
        * `avg_work_hours` = mean `work_hours` over records that are present-status **and** have a `work_hours` value
          (average over records, not an average of per-employee averages); `null` if there are none. R8 rounding.
        * `late_count` = records with `late_minutes` > 0; `total_late_minutes` = their sum.
        * `leave_count` = `LEAVE` records; `on_duty_count` = `ON_DUTY` records.
        A department whose headcount is 0 for the month is omitted. Department comes from the `employees` collection.
      parameters:
        - {$ref: '#/components/parameters/Month'}
        - name: department
          in: query
          required: false
          schema: {type: string}
          description: Restrict to one department (exact match).
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DepartmentSummary'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /analytics/leaderboard/late:
    get:
      tags: [analytics]
      operationId: lateLeaderboard
      summary: Late-comers leaderboard for a month
      description: |
        Ranks employees by total `late_minutes` in the month (only employees with a total > 0), highest first.
        * **Ranking is standard competition ranking** - ties share a rank and the next rank is skipped
          (e.g. 1, 2, 2, 4). Do not use the array position as the rank.
        * `limit` applies **after** ranking: return every employee whose `rank` <= `limit`. The response can therefore
          contain more than `limit` rows when employees tie at a rank <= `limit`.
        * Order of rows: `total_late_minutes` descending, then `emp_code` ascending.
        * When `department` is given, ranking happens inside that department only.
        `name` and `department` come from `employees`; logs whose employee does not exist are ignored.
      parameters:
        - {$ref: '#/components/parameters/Month'}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 50, default: 10}}
        - {name: department, in: query, required: false, schema: {type: string}}
      responses:
        '200':
          description: Leaderboard
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Leaderboard'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /analytics/departments/{department}/trend:
    get:
      tags: [analytics]
      operationId: departmentTrend
      summary: Daily attendance trend for a department
      description: |
        Returns **exactly one row per calendar day** from `from` to `to` inclusive - weekends and days with no records
        at all included (use a gap-filling stage such as `$densify`; do not loop in Python).
        * `is_working_day` = Mon-Fri.
        * `headcount` = department employees with `joined_on` <= that day (R9).
        * `present_count` = presence-status records that day (half day = 0.5), any weekday or weekend.
        * `late_count` = records that day with `late_minutes` > 0.
        * `attendance_rate` = `present_count` / `headcount`, 4 decimals half-up, **only on working days with headcount > 0,
          otherwise `null`**. A working day with no records has rate 0.0.
        * `moving_avg_7d` = mean of the non-null `attendance_rate` values among the current row and up to 6 preceding rows
          **inside the requested range** (rows before `from` are not considered, so the first rows use a shorter
          window). `null` if the window holds no non-null rate. 4 decimals half-up.
        Errors: 404 unknown department (no employees); 422 if `to` < `from` or the range exceeds 92 days.
      parameters:
        - {name: department, in: path, required: true, schema: {type: string}}
        - {name: from, in: query, required: true, schema: {type: string, format: date}}
        - {name: to, in: query, required: true, schema: {type: string, format: date}}
      responses:
        '200':
          description: Trend
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Trend'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /admin/explain/{endpoint}:
    get:
      tags: [admin]
      operationId: explainEndpoint
      summary: MongoDB explain for an endpoint's main query
      description: |
        Runs `explain` with `executionStats` verbosity for the **main query or aggregation of the named endpoint,
        built exactly as that endpoint builds it** (same filters, sort, pagination), and returns the raw explain output.
        Accepts the same query parameters as the endpoint it explains (for `department_trend` pass `department`,
        `from`, `to`; for `employee_monthly` pass `emp_code`, `month`; etc.). Missing required parameters -> 422.
        The grader checks that the plan contains `IXSCAN` and no `COLLSCAN`, on a ~100k-record dataset.
      parameters:
        - name: endpoint
          in: path
          required: true
          schema:
            type: string
            enum: [attendance_list, employee_monthly, department_summary, late_leaderboard, department_trend]
        - {name: emp_code, in: query, required: false, schema: {type: string}}
        - {name: month, in: query, required: false, schema: {type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$'}}
        - {name: department, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 50, default: 10}}
        - {name: date_from, in: query, required: false, schema: {type: string, format: date}}
        - {name: date_to, in: query, required: false, schema: {type: string, format: date}}
        - {name: status, in: query, required: false, schema: {$ref: '#/components/schemas/Status'}}
        - {name: from, in: query, required: false, schema: {type: string, format: date}}
        - {name: to, in: query, required: false, schema: {type: string, format: date}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200':
          description: Explain output
          content:
            application/json:
              schema:
                type: object
                required: [endpoint, collection, explain]
                properties:
                  endpoint: {type: string}
                  collection: {type: string, description: Collection the explained operation ran on.}
                  explain: {type: object, additionalProperties: true, description: Raw MongoDB explain document as JSON.}
        '422': {$ref: '#/components/responses/ValidationError'}
components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Month:
      name: month
      in: query
      required: true
      description: Calendar month, `YYYY-MM`.
      schema: {type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$'}
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [detail]
            properties:
              detail: {type: string}
    ValidationError:
      description: Validation error (FastAPI default body)
      content:
        application/json:
          schema:
            type: object
            required: [detail]
            properties:
              detail: {type: array, items: {type: object, additionalProperties: true}}
  schemas:
...
    Status:
      type: string
      enum: [PRESENT, ABSENT, LEAVE, WFH, ON_DUTY]
      description: See the status table in the API description.
...
    EmployeeMonthly:
      type: object
      required: [emp_code, month, working_days, present_days, leave_days, late_count, total_late_minutes, total_overtime_minutes, attendance_pct]
      properties:
        emp_code: {type: string}
        month: {type: string}
        working_days: {type: integer}
        present_days: {type: number}
        leave_days: {type: integer}
        late_count: {type: integer}
        total_late_minutes: {type: integer}
        total_overtime_minutes: {type: integer}
        attendance_pct: {type: number, nullable: true}
    DepartmentSummary:
      type: object
      required: [month, items]
      properties:
        month: {type: string}
        items:
          type: array
          items:
            type: object
            required: [department, headcount, present_days, avg_work_hours, late_count, total_late_minutes, leave_count, on_duty_count]
            properties:
              department: {type: string}
              headcount: {type: integer}
              present_days: {type: number}
              avg_work_hours: {type: number, nullable: true}
              late_count: {type: integer}
              total_late_minutes: {type: integer}
              leave_count: {type: integer}
              on_duty_count: {type: integer}
    Leaderboard:
      type: object
      required: [month, items]
      properties:
        month: {type: string}
        items:
          type: array
          items:
            type: object
            required: [rank, emp_code, name, department, total_late_minutes, late_count]
            properties:
              rank: {type: integer, minimum: 1}
              emp_code: {type: string}
              name: {type: string}
              department: {type: string}
              total_late_minutes: {type: integer}
              late_count: {type: integer}
    Trend:
      type: object
      required: [department, items]
      properties:
        department: {type: string}
        items:
          type: array
          items:
            type: object
            required: [date, is_working_day, headcount, present_count, late_count, attendance_rate, moving_avg_7d]
            properties:
              date: {type: string, format: date}
              is_working_day: {type: boolean}
              headcount: {type: integer}
              present_count: {type: number}
              late_count: {type: integer}
              attendance_rate: {type: number, nullable: true}
              moving_avg_7d: {type: number, nullable: true}